requirements.txt
# Core ML Libraries
scikit-learn==1.3.2
lightgbm==4.1.0
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
torch>=1.9.0
transformers>=4.11.0

# Technical Analysis and Market Data
ta==0.10.2
yfinance==0.2.28
ccxt>=1.60.0
binance-connector>=1.3.0
talib-binary>=0.4.19

# API and Web
flask==3.0.0
flask-cors==4.0.0
websockets==12.0
requests==2.31.0
aiohttp>=3.8.0

# Sentiment Analysis
textblob>=0.15.3
vaderSentiment>=3.3.2
tweepy>=4.4.0
praw>=7.4.0
newsapi-python>=0.2.6

# Statistical Models and Time Series
statsmodels==0.14.1
prophet>=1.0.1
arch>=4.19.0  # For GARCH models

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
optuna==3.5.0
loguru==0.7.2

# Data Processing
imbalanced-learn==0.11.0

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1

# Development
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910